+++ src/demo/views/Components.vue
@@ -21,6 +21,7 @@
     <Avatar class="mt-10" />
     <NumberInputs class="mt-10" />
     <DatePickerSection class="mt-10" />
+    <TimePickerSection class="mt-10" />
     <Selects class="mt-10" />
     <TextAreas class="mt-10" />
     <Toasts class="mt-10" />
@@ -36,6 +37,7 @@ import Buttons from '../sections/Buttons.vue'
 import Cards from '../sections/Cards.vue'
 import Checkboxes from '../sections/Checkboxes.vue'
 import DatePickerSection from '../sections/DatePickerSection.vue'
+import TimePickerSection from '../sections/TimePickerSection.vue'
 import Inputs from '../sections/Inputs.vue'
 import Links from '../sections/Links.vue'
 import Popover from '../sections/Popover.vue'
@@ -61,6 +63,7 @@ import Loader from '../sections/Loader.vue'
     Cards,
     Checkboxes,
     DatePickerSection,
+    TimePickerSection,
     Inputs,
     Links,
     NumberInputs,